%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{ACNE-package}
\alias{ACNE-package}
\alias{ACNE}
\docType{package}

\title{Package ACNE}


\description{
  A summarization method to estimate allele-specific copy number signals for Affymetrix SNP microarrays using non-negative matrix factorization (NMF).
}

\section{Installation and updates}{
  This package requires the \pkg{aroma.affymetrix} package.
  To install this package, do:
  \code{install.packages("ACNE")}
}

\section{To get started}{
 \enumerate{
  \item For a one-command pipeline, see the \code{\link{doACNE}}() method.
  \item For other usages, see the \code{\link{NmfPlm}} class.
 }
}

\author{Maria Ortiz, Henrik Bengtsson, Angel Rubio}

\section{License}{
  LGPL (>= 2.1)
}

\references{
 [1] M. Ortiz-Estevez, H. Bengtsson, A. Rubio, \emph{ACNE: a summarization method to estimate allele-specific copy numbers for Affymetrix SNP arrays}, Bioinformatics, 2010 [PMC2913655].
 \cr
}
\keyword{package}
