%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  pseudoinverse.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{pseudoinverse}
\alias{pseudoinverse}


\title{Calculates the pseudo inverse of a matrix}

\description{
 Calculates the pseudo inverse of a matrix.
}

\usage{
pseudoinverse(X, ...)
}

\arguments{
 \item{X}{A \code{\link[base]{numeric}} \code{\link[base]{matrix}}.}
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{matrix}}.
}

\seealso{
  Internally \code{\link[base]{svd}}() is used.
}


\keyword{internal}
