%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{ACNE-package}
\alias{ACNE-package}
\alias{ACNE}
\docType{package}

\title{Package ACNE}


 \description{
   Package for NMF summarization of SNP probes.

   This package should be considered to be in an alpha phase.
   You should expect the API to be changing over time.
 }

 \section{Requirements}{
   This package requires the \pkg{aroma.affymetrix} package.
 }

 \section{Installation and updates}{
   To install this package, see instructions at
   \url{http://r-forge.r-project.org/projects/snpsprocessing/}.
 }

 \section{To get started}{
   See the \code{\link{NmfPlm}} class.
 }

 \author{
   Angel Rubio <arubio@ceit.es>,
   Maria Ortiz <mortiz@ceit.es>,
   Henrik Bengtsson <henrikb@braju.com>
 }

 \section{License}{
   LGPL (>= 2.1)
 }

 \references{
   TBA.
 }
\keyword{package}
